+++ lib/ig_markets/limited_risk_premium.rb
@@ -1,0 +1,8 @@
+module IGMarkets
+  # Contains details on a limited risk premium. Returned by {Instrument#limited_risk_premium{,
+  # {Position#limited_risk_premium} and {WorkingOrder#limited_risk_premium}.
+  class LimitedRiskPremium < Model
+    attribute :unit, Symbol, allowed_values: [:percentage, :points]
+    attribute :value, Float
+  end
+end