+++ spec/factories/ig_markets/limited_risk_premium.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  factory :limited_risk_premium, class: IGMarkets::LimitedRiskPremium do
+    unit 'POINTS'
+    value 10
+  end
+end