+++ spec/factories/ig_markets/position.rb
@@ -10,6 +10,7 @@ FactoryGirl.define do
     direction 'BUY'
     level 100.0
     limit_level 110.0
+    limited_risk_premium { build :limited_risk_premium }
     size 10.4
     stop_level 90.0
     trailing_step nil